Rewritten data preprocessing scripts Added PatchDock runner Changed alignment logic

+++ src/main/java/ru/ifmo/docking/benchmark/PatchdockRunner.java
@@ -13,7 +13,7 @@ import java.io.PrintWriter;
 import java.util.List;
 import java.util.function.Supplier;
 
-public class ZdockRunner implements DockerRunner {
+public class PatchdockRunner implements DockerRunner {
 
     @Override
     public List<Supplier<Protein>> run(String complex, File benchmarkDir, Timer timer) {
@@ -32,10 +32,7 @@ public class ZdockRunner implements DockerRunner {
         File unboundReceptorPdb = new File(complexDir, unboundReceptorFile + ".pdb");
         File unboundLigandPdb = new File(complexDir, unboundLigandFile + ".pdb");
 
-        System.out.println("Preprocessing receptor");
-        preprocess(unboundReceptorPdb);
-        System.out.println("Preprocessing ligand");
-        preprocess(unboundLigandPdb);
+        preprocess(unboundReceptorPdb, unboundLigandPdb);
 
         File resultTs = new File(resultsFileDir, complex + ".ts");
 
@@ -46,26 +43,18 @@ public class ZdockRunner implements DockerRunner {
         } else {
             System.out.println("Start docking");
             timer.start();
-            runProcess(
-                    "mpirun",
-                    "-np",
-                    "" + Runtime.getRuntime().availableProcessors(),
-                    "zdock",
-                    "-N", "50000",
-                    "-o", complex + ".out",
-                    "-R", unboundReceptorFile + "_m.pdb",
-                    "-L", unboundLigandFile + "_m.pdb");
+            runProcess("./patch_dock.Linux", "params.txt", complex + ".out");
             timer.stop();
             writeLong(resultTs, timer.getTime());
             runProcess("cp", complex + ".out", resultsFileDir.getName() + "/" + complex + ".out");
         }
 
         System.out.println("Start writing results");
-        runProcess("./create.pl", complex + ".out", "50000");
+        runProcess("./transOutput.pl", complex + ".out", "0", "500000");
 
         List<Supplier<Protein>> results = Lists.newArrayList();
         File currentDir = new File(".");
-        File[] files = currentDir.listFiles((dir, name) -> name.startsWith("complex."));
+        File[] files = currentDir.listFiles((dir, name) -> name.startsWith(complex + ".") && name.endsWith(".pdb"));
         for (File file : files) {
             results.add(() -> PdbUtil.readSimplifiedPdbFile(file));
         }
@@ -74,11 +63,10 @@ public class ZdockRunner implements DockerRunner {
 
     @Override
     public void cleanup(String complex) {
-        deleteFile(new File(complex + "_r_u_m.pdb"));
-        deleteFile(new File(complex + "_l_u_m.pdb"));
+        deleteFile(new File("params.txt"));
         deleteFile(new File(complex + ".out"));
         File currentDir = new File(".");
-        File[] files = currentDir.listFiles((dir, name) -> name.startsWith("complex."));
+        File[] files = currentDir.listFiles((dir, name) -> name.startsWith(complex + ".") && name.endsWith(".pdb"));
         for (File file : files) {
             deleteFile(file);
         }
@@ -86,18 +74,20 @@ public class ZdockRunner implements DockerRunner {
 
     @Override
     public String getName() {
-        return "zdock";
+        return "patchdock";
     }
 
     private void deleteFile(File file) {
-        if (!file.delete()) {
-            runProcess("rm", file.getAbsolutePath());
+        if (file.exists()) {
+            if (!file.delete()) {
+                runProcess("rm", file.getAbsolutePath());
+            }
         }
     }
 
-    private void preprocess(File pdb) {
+    private void preprocess(File receptorPdb, File ligandPdb) {
         try {
-            runProcess("./mark_sur", pdb.getCanonicalPath(), pdb.getName().replace(".pdb", "_m.pdb"));
+            runProcess("./buildParams.pl", receptorPdb.getCanonicalPath(), ligandPdb.getCanonicalPath(), "4.0");
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
